+++ spd/src/lopicost/spd/robot/bean/rd/Toma.java
@@ -4,37 +4,19 @@ package lopicost.spd.robot.bean.rd;
 
 public class Toma {
 	private int ordenToma;
-	private int cantidadToma;
+	private double cantidadToma;
 	private String idToma;
 	private String nombreToma;
 	
 	
-	public int getOrdenToma() {
-		return ordenToma;
-	}
-	public void setOrdenToma(int ordenToma) {
-		this.ordenToma = ordenToma;
-	}
-	public int getCantidadToma() {
-		return cantidadToma;
-	}
-	public void setCantidadToma(int cantidadDia) {
-		this.cantidadToma = cantidadDia;
-	}
-	public String getIdToma() {
-		return idToma;
-	}
-	public void setIdToma(String idToma) {
-		this.idToma = idToma;
-	}
-	public String getNombreToma() {
-		return nombreToma;
-	}
-	public void setNombreToma(String nombreToma) {
-		this.nombreToma = nombreToma;
-	}
-	
-	
+	public int getOrdenToma() 						{		return ordenToma;					}
+	public void setOrdenToma(int ordenToma) 		{		this.ordenToma = ordenToma;			}
+	public double getCantidadToma() 				{		return cantidadToma;				}
+	public void setCantidadToma(double cantidadDia) {		this.cantidadToma = cantidadDia;	}
+	public String getIdToma() 						{		return idToma;						}
+	public void setIdToma(String idToma) 			{		this.idToma = idToma;				}
+	public String getNombreToma() 					{		return nombreToma;					}
+	public void setNombreToma(String nombreToma) 	{		this.nombreToma = nombreToma;		}
 	
 	
 }
